More execution processing refactors
