+++ src/Microsoft.SqlTools.ServiceLayer/Management/Common/ExecutionMode.cs
@@ -5,7 +5,7 @@
 
 using System;
 
-namespace Microsoft.SqlTools.ServiceLayer.Agent
+namespace Microsoft.SqlTools.ServiceLayer.Management
 {
     /// <summary>
     /// Execution mode enumeration Success if execution succeeded of Failure otherwise for now.
