+++ src/Microsoft.SqlTools.ServiceLayer/Management/Common/ISqlControlCollection.cs
@@ -5,7 +5,7 @@
 
 using System;
 
-namespace Microsoft.SqlTools.ServiceLayer.Agent
+namespace Microsoft.SqlTools.ServiceLayer.Management
 {
 	/// <summary>
 	/// defines notion of sitable object
