+++ src/Microsoft.SqlTools.ServiceLayer/Management/Common/ManagementActionBase.cs
@@ -9,28 +9,23 @@ using System.Collections.Specialized;
 using System.Data;
 using System.Diagnostics;
 using System.IO;
-using System.Xml;
 using System.Threading;
+using System.Text;
+using System.Xml;
 using Microsoft.SqlServer.Management.Common;
 using Microsoft.SqlServer.Management.Diagnostics;
 using Microsoft.SqlServer.Management.Smo;
 using Microsoft.SqlTools.ServiceLayer.Admin;
 using Microsoft.SqlTools.ServiceLayer.Agent;
-using System.Text;
 
 namespace Microsoft.SqlTools.ServiceLayer.Management
 {
     /// <summary>
-    /// base class that can be used to derived from for the main classes [containers] of the dialogs
+    /// base class that can be used to derived from for the main classes
     /// </summary>
-    public class ManagementActionBase : IDisposable, IExecutionAwareManagementAction, IManagementAction
+    public class ManagementActionBase : IDisposable, IExecutionAwareManagementAction
     {
 #region Members
-
-        /// <summary>
-        /// selected node as specified to SelectNode method
-        /// </summary>
-        //private TreeNode    selectedNode;
 
         /// <summary>
         /// service provider of our host. We should direct all host-specific requests to the services
@@ -141,10 +136,10 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
         /// </returns>
         public bool PreProcessExecution(PreProcessExecutionInfo executionInfo, out ExecutionMode executionResult)
         {
-            //we start from failure
+            // we start from failure
             executionResult = ExecutionMode.Failure;
 
-            //OK, we do server switching for scripting for SQL/Analysis Server execution here
+            // OK, we do server switching for scripting for SQL Server execution here
             RunType runType = executionInfo.RunType;
             if (IsScripting(runType))
             {
@@ -156,15 +151,13 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
 
             if (DataContainer != null)
             {
-                //we take over execution here. We substitute the server here for AMO and SQL
-                //dialogs
+                // we take over execution here. We substitute the server here for SQL containers
                 if (DataContainer.ContainerServerType == CDataContainer.ServerType.SQL)
                 {                
                     ExecuteForSql(executionInfo, out executionResult);
                     return false;//execution of the entire control was done here
                 }
             }
-
 
             // call virtual function to do regular execution
             return DoPreProcessExecution(executionInfo.RunType, out executionResult);
@@ -180,6 +173,20 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
         {
             //delegate to the protected virtual method
             DoPostProcessExecution(runType, executionResult);
+        }
+
+        /// <summary>
+        /// called when the host received Cancel request. NOTE: this method can return while
+        /// operation is still being canceled
+        /// </summary>
+        /// <returns>
+        /// true if the host should do standard cancel for the currently running view or
+        /// false if the Cancel operation was done entirely inside this method and there is nothing
+        /// extra that should be done
+        /// </returns>
+        public bool Cancel()
+        {
+            return true;
         }
 #endregion
 
@@ -344,7 +351,7 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
         }
 
         /// <summary>
-        /// determines whether we need to substitute SMO/AMO server objects with the
+        /// determines whether we need to substitute SMO server objects with the
         /// temporary ones while doing scripting
         /// </summary>
         /// <returns></returns>
@@ -405,25 +412,24 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
             }
             catch (System.Exception)
             {
-                //We may not have a valid dialog subject here (such as if the object hasn't been created yet)
-                //so in that case we'll just ignore it as that's a normal scenario. 
+                // We may not have a valid dialog subject here (such as if the object hasn't been created yet)
+                // so in that case we'll just ignore it as that's a normal scenario. 
             }
 
             StringCollection sc = GetServerConnectionForScript().CapturedSql.Text;
-            //Scripting may happen on either the server ExecutionManager or the
-            //ExecutionManager of the object itself. So we make sure to check
-            //the subject text if the server ExecutionManager didn't have any
-            //scripts after doing the scripting
+            // Scripting may happen on either the server ExecutionManager or the
+            // ExecutionManager of the object itself. So we make sure to check
+            // the subject text if the server ExecutionManager didn't have any
+            // scripts after doing the scripting
             if (sc.Count == 0 && sqlDialogSubject != null)
             {
                 sc = sqlDialogSubject.ExecutionManager.ConnectionContext.CapturedSql.Text;
             }
-            int                 i;
-            StringBuilder       script  = new StringBuilder(4096);  
 
+            StringBuilder script  = new StringBuilder(4096);
             if (sc != null)
             {
-                for (i = 0; i < sc.Count; i ++)
+                for (int i = 0; i < sc.Count; i ++)
                 {
                     script.AppendFormat("{0}\r\nGO\r\n", sc[i].ToString());
                 }
@@ -432,9 +438,8 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
             return script.ToString();
         }
 
-
         /// <summary>
-        /// called when we need to script a Sql server dlg.
+        /// called when we need to script a Sql Server action
         /// </summary>
         /// <param name="executionInfo"></param>
         /// <param name="executionResult"></param>
@@ -452,13 +457,13 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
                 DataContainer.Server = new Microsoft.SqlServer.Management.Smo.Server(DataContainer.ServerConnection);
             }
 
-            String szScript = null;
+            string szScript = null;
             bool isScripting = IsScripting(executionInfo.RunType);
             var executionModeOriginal = GetServerConnectionForScript().SqlExecutionModes;
-            //For Azure the ExecutionManager is different depending on which ExecutionManager
-            //used - one at the Server level and one at the Database level. So to ensure we
-            //don't use the wrong execution mode we need to set the mode for both (for on-prem
-            //this will essentially be a no-op)
+            // For Azure the ExecutionManager is different depending on which ExecutionManager
+            // used - one at the Server level and one at the Database level. So to ensure we
+            // don't use the wrong execution mode we need to set the mode for both (for on-prem
+            // this will essentially be a no-op)
             SqlExecutionModes subjectExecutionModeOriginal = executionModeOriginal;
             SqlSmoObject sqlDialogSubject = null;
             try
@@ -467,8 +472,8 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
             }
             catch (System.Exception)
             {
-                //We may not have a valid dialog subject here (such as if the object hasn't been created yet)
-                //so in that case we'll just ignore it as that's a normal scenario. 
+                // We may not have a valid dialog subject here (such as if the object hasn't been created yet)
+                // so in that case we'll just ignore it as that's a normal scenario. 
             }
 
             if (sqlDialogSubject != null)
@@ -480,7 +485,7 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
                 SqlExecutionModes newMode = isScripting
                     ? SqlExecutionModes.CaptureSql
                     : SqlExecutionModes.ExecuteSql;
-                //now, do the execution itself
+                // now, do the execution itself
                 GetServerConnectionForScript().SqlExecutionModes = newMode;
                 if (sqlDialogSubject != null)
                 {
@@ -639,20 +644,18 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
 //             canScriptToWindow = canScriptToFile = canScriptToClipboard = canScriptToJob = true;
 //         }
 // #endregion
-//         protected IServiceProvider ServiceProvider
-//         {
-//             get
-//             {
-//                 if (this.serviceProvider == null)
-//                 {
-//                     STrace.Assert(false, "Cannot work without service provider!");
-//                     STrace.LogExThrow();
-//                     //BUGBUG - should we have our own exception here?
-//                     throw new InvalidOperationException();
-//                 }
-//                 return this.serviceProvider;
-//             }
-//         }
+        protected IServiceProvider ServiceProvider
+        {
+            get
+            {
+                if (this.serviceProvider == null)
+                {
+
+                    throw new InvalidOperationException();
+                }
+                return this.serviceProvider;
+            }
+        }
 //         /// <summary>
 //         /// returns combination of the given 2 arrays
 //         /// </summary>
@@ -678,11 +681,6 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
 //             }
 //         }
 
-////////////////////////////////////
-/// 
-/// 
-
-
         /// <summary>
         /// execution mode by default for now is success
         /// </summary>
@@ -703,14 +701,13 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
             }
         }
 
-         public virtual ExecutionMode LastExecutionResult
+        public virtual ExecutionMode LastExecutionResult
         {
             get
             {
                 return ExecutionMode;
             }
         }
-
 
         /// <summary>
         /// Overridable function that allow a derived class to implement
@@ -735,60 +732,9 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
 
         /// <summary>
         /// Overridable function that allow a derived class to implement its
-        /// OnReset functionality
+        /// OnScript functionality. 
         /// </summary>
         /// <param name="sender"></param>
-        // public virtual void OnReset(object sender)
-        // {
-        //     //nothing
-        // }
-
-
-        // /// <summary>
-        // /// Overridable function that allow a derived class to implement its
-        // /// OnScriptToWindow functionality
-        // /// </summary>
-        // /// <param name="sender"></param>
-        // public virtual string OnScriptToWindow(object sender)
-        // {
-        //     //redirect to the single scripting virtual method by default
-        //     return Script();
-        // }
-
-        // /// <summary>
-        // /// Overridable function that allow a derived class to implement its
-        // /// OnScriptToWindow functionality
-        // /// </summary>
-        // /// <param name="sender"></param>
-        // public virtual string OnScriptToFile(object sender)
-        // {
-        //     //redirect to the single scripting virtual method by default
-        //     return Script();
-        // }
-
-
-        // /// <summary>
-        // /// Overridable function that allow a derived class to implement its
-        // /// OnScriptToClipboard functionality. 
-        // /// </summary>
-        // /// <param name="sender"></param>
-        // public virtual string OnScriptToClipboard(object sender)
-        // {
-        //     //redirect to the single scripting virtual method by default
-        //     return Script();
-        // }
-
-        // /// <summary>
-        // /// Overridable function that allow a derived class to implement its
-        // /// OnScriptToJob functionality. 
-        // /// </summary>
-        // /// <param name="sender"></param>
-        // public virtual string OnScriptToJob(object sender)
-        // {
-        //     //redirect to the single scripting virtual method by default
-        //     return Script();
-        // }
-
         public virtual string OnScript(object sender)
         {
             //redirect to the single scripting virtual method by default
@@ -803,12 +749,12 @@ namespace Microsoft.SqlTools.ServiceLayer.Management
         /// <returns></returns>
         protected virtual string Script()
         {
-            //redirect to the RunNow method. Our host should be turning script capture on and off for
-            //OLAP/SQL servers and composing the text of the resulting script by itself
+            // redirect to the RunNow method. Our host should be turning script capture on and off for
+            // OLAP/SQL servers and composing the text of the resulting script by itself
             OnRunNow(this);
 
-            //null is a special value. It means that we want to indicate that we didn't want to generate 
-            //script text
+            // null is a special value. It means that we want to indicate that we didn't want to generate 
+            // script text
             return null;
         }
