+++ src/Microsoft.SqlTools.ServiceLayer/Management/Common/ProgressReportCommon.cs
@@ -8,7 +8,7 @@ using System.Drawing;
 using System.Threading;
 using System.Runtime.InteropServices;
 
-namespace Microsoft.SqlTools.ServiceLayer.Agent
+namespace Microsoft.SqlTools.ServiceLayer.Management
 {
 	/// <summary>
 	/// Enumeration for status of individual actions
