+++ src/Microsoft.SqlTools.ServiceLayer/Management/Common/RunType.cs
@@ -5,7 +5,7 @@
 
 using System;
 
-namespace Microsoft.SqlTools.ServiceLayer.Agent
+namespace Microsoft.SqlTools.ServiceLayer.Management
 {
     /// <summary>
     /// what type of actions does the worker know to execute
@@ -18,6 +18,5 @@ namespace Microsoft.SqlTools.ServiceLayer.Agent
         ScriptToWindow,
         ScriptToClipboard,
         ScriptToJob
-    }
-    
+    }   
 }
