+++ test/Microsoft.SqlTools.ServiceLayer.IntegrationTests/Agent/AgentJobTests.cs
@@ -30,7 +30,27 @@ namespace Microsoft.SqlTools.ServiceLayer.IntegrationTests.Agent
                 var connectionResult = await LiveConnectionHelper.InitLiveConnectionInfoAsync("master", queryTempFile.FilePath);
                 await service.HandleCreateAgentJobRequest(new CreateAgentJobParams
                 {
-                    OwnerUri = connectionResult.ConnectionInfo.OwnerUri
+                    OwnerUri = connectionResult.ConnectionInfo.OwnerUri,
+                    Job = new AgentJobInfo()
+                    {
+                        Name = "Test Job",
+                        Owner = "sa",
+                        Description = "Test job description",
+                        CurrentExecutionStatus = 1,
+                        LastRunOutcome = 1,
+                        CurrentExecutionStep = "Step 1",
+                        Enabled = false,
+                        HasTarget = false,
+                        HasSchedule = false,
+                        HasStep = false,
+                        Runnable = true,
+                        Category = "Cateory 1",
+                        CategoryId = 1,
+                        CategoryType = 1,
+                        LastRun = "today",
+                        NextRun = "tomorrow",
+                        JobId = "Job1"
+                    }
                 }, createContext.Object);
                 createContext.VerifyAll();
             }
